Simplified the codebase
