+++ .env.example
@@ -4,10 +4,6 @@
 LOGIN = 'example@example.com'
 PASSWORD = 'password'
 
-# How to bypass Cloudflare. cloudproxy is recommended, but can be cloudscraper or requests.
-REQUEST_METHOD = 'cloudproxy'
-
-# Used only if you use CloudProxy, default is if you run it on the same machine
 CLOUDPROXY_URL = 'http://localhost:8191/v1'
 
 OUTPUT_FOLDER = 'plugins'
