+++ .vscode/settings.json
@@ -1,3 +1,5 @@
 {
-    "python.pythonPath": "env/bin/python"
+    "python.pythonPath": "env/bin/python",
+    "python.linting.pylintEnabled": true,
+    "python.linting.enabled": true
 }
